ref(py): Move the CacheReader to utils
